+++ libs/cardano-ledger-core/src/Cardano/Ledger/Plutus/CostModels.hs
@@ -134,17 +134,24 @@ instance NFData CostModel where
   rnf (CostModel lang cm ectx) = lang `deepseq` cm `deepseq` rnf ectx
 
 instance FromJSON CostModels where
-  parseJSON = parseCostModels True
+  parseJSON = parseCostModels True []
 
 parseCostModels ::
   -- | Do not restrict number of parameters to the initial count and allow parsing of cost models
   -- for unknown plutus versions.
   Bool ->
+  -- | Restrict parsable Plutus language versions to the given list.
+  -- If left empty, no restrictions are applied and all non-native languages
+  -- are parsed.
+  [Language] ->
   Value ->
   Parser CostModels
-parseCostModels isLenient =
+parseCostModels isLenient languages =
   withObject "CostModels" $ \o -> do
-    cms <- mapM (parseCostModel isLenient o) nonNativeLanguages
+    cms <-
+      if null languages
+        then mapM (parseCostModel isLenient o) nonNativeLanguages
+        else mapM (parseCostModel isLenient o) languages
     let cmsMap = Map.fromList [(cmLanguage cm, cm) | Just cm <- cms]
     unknownCostModels <-
       if isLenient